refactor(inputConsole): 입력 예외사항 메서드 이동
