+++ src/main/java/car/domain/Car.java
@@ -1,10 +1,11 @@
 package car.domain;
 
 import static car.option.Option.INITIAL_POSITION;
-import static car.option.Option.MAX_NAME_LENGTH;
-import static car.option.Option.MIN_NAME_LENGTH;
 
-public class Car {
+public class
+
+
+Car {
 
     private final String name;
     private int position;
@@ -14,10 +15,15 @@ public class Car {
     }
 
     public Car(String name, int position) {
-        validateNameLength(name);
         validatePositionOverInitialPosition(position);
         this.name = name;
         this.position = position;
+    }
+
+    private void validatePositionOverInitialPosition(int position) {
+        if (position < INITIAL_POSITION) {
+            throw new IllegalArgumentException("[ERROR] 위치는 시작점보다 작으면 안됩니다.");
+        }
     }
 
     public int selectMaxPosition(int otherPosition) {
@@ -34,20 +40,6 @@ public class Car {
         }
     }
 
-    private void validateNameLength(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("[ERROR] 이름이 너무 깁니다.");
-        }
-        if (name.length() < MIN_NAME_LENGTH) {
-            throw new IllegalArgumentException("[ERROR] 이름이 너무 짧습니다.");
-        }
-    }
-
-    private void validatePositionOverInitialPosition(int position) {
-        if (position < INITIAL_POSITION) {
-            throw new IllegalArgumentException("[ERROR] 위치는 시작점보다 작으면 안됩니다.");
-        }
-    }
 
     public String getName() {
         return name;
