+++ src/main/java/car/view/InputView.java
@@ -4,8 +4,6 @@ import java.util.Scanner;
 import java.util.regex.Pattern;
 
 public class InputView {
-//    private static final Scanner scanner = new Scanner(System.in);
-//    private static final Pattern NOT_INTEGER_PATTERN = Pattern.compile("\\D+");
     private final Input input;
     public InputView(Input input){
         this.input = input;
@@ -17,21 +15,6 @@ public class InputView {
 
     public int inputTrialCount() {
         System.out.println("시도할 회수는 몇회인가요?");
-//        String input = scanner.nextLine();
-//        validateInteger(input);
-//        validateTrialCountUnderMaxInteger(input);
         return Integer.parseInt(input.inputTrialCount());
     }
-
-//    private void validateInteger(String input) {
-//        if (NOT_INTEGER_PATTERN.matcher(input).matches()) {
-//            throw new IllegalArgumentException("시도 횟수는 정수여야 합니다");
-//        }
-//    }
-//
-//    private void validateTrialCountUnderMaxInteger(String input) {
-//        if (input.length() >= MAX_TRIAL_COUNT_LENGTH) {
-//            throw new IllegalArgumentException("시도 횟수는 999999999이하여야 합니다.");
-//        }
-//    }
 }
